Add files via upload
This is the first-reworked version of the program! The most important update is related to that of the maze.py since it now has a section for buttons. Of course, all that was added is, for the time being, a place holder.

This update is old, but since had to reconstruct this git for an error I made, it will be uploaded during this time.